requirements.txt
gspread
Flask==2.0.1
google-api-python-client==2.2.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
gunicorn==20.1.0
Flask
python-dotenv